server/requirements.txt
asgiref==3.2.3
Django==3.0.2
djangorestframework==3.11.0
Markdown==3.1.1
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
